Split basic.cpp into 2 files
